src/config.py
"""Configuration helpers for benchmark runs."""

from __future__ import annotations

import json
import os
from dataclasses import asdict, dataclass, field
from pathlib import Path
from typing import Any

try:
    import yaml
except ModuleNotFoundError:
    yaml = None


def _expand_path(value: str | None) -> str | None:
    if not value:
        return value
    return os.path.expandvars(os.path.expanduser(value))


@dataclass
class ProviderConfig:
    name: str = "ollama"
    model: str | None = None
    embedding_model: str | None = None
    rerank_model: str | None = None
    api_key: str | None = None
    api_key_env: str | None = None
    base_url: str | None = None
    api_version: str | None = None
    aws_region: str | None = None
    project_id: str | None = None
    location: str | None = None
    headers: dict[str, str] = field(default_factory=dict)
    request_template: dict[str, Any] = field(default_factory=dict)
    response_text_path: str | None = None
    response_tokens_path: str | None = None
    timeout_seconds: float = 60.0
    max_retries: int = 2
    retry_backoff_seconds: float = 1.5
    temperature: float = 0.2
    max_tokens: int = 2048


@dataclass
class JudgeConfig:
    strategy: str = "auto"
    ollama_model: str = "qwen3.5:9b"
    timeout_seconds: float = 30.0
    pass_threshold: float = 0.7
    review_threshold: float = 0.4
    attacker_model: str = "qwen3.5:9b"
    attacker_base_url: str = "http://localhost:11434"
    attacker_api_key: str | None = None
    attacker_api_key_env: str | None = None
    attacker_temperature: float = 0.3
    pattern_detectors_enabled: bool = True
    openai_moderation_enabled: bool = False
    openai_moderation_model: str = "omni-moderation-latest"
    openai_moderation_api_key: str | None = None
    openai_moderation_api_key_env: str | None = None
    openai_moderation_base_url: str | None = None
    perspective_enabled: bool = False
    perspective_api_key: str | None = None
    perspective_api_key_env: str | None = None
    perspective_threshold: float = 0.7
    harmjudge_enabled: bool = False
    harmjudge_model: str | None = None
    harmjudge_api_key: str | None = None
    harmjudge_api_key_env: str | None = None
    harmjudge_base_url: str | None = None
    external_detector_url: str | None = None
    external_detector_api_key: str | None = None
    external_detector_api_key_env: str | None = None
    external_detector_headers: dict[str, str] = field(default_factory=dict)
    external_detector_timeout_seconds: float = 15.0
    detector_weight: float = 0.25
    detector_family_enabled: dict[str, bool] = field(default_factory=dict)
    detector_family_weights: dict[str, float] = field(default_factory=dict)


@dataclass
class BenchmarkConfig:
    prompt_file: str = "prompts/customer-support-bot.txt"
    test_file: str = "tests/safeprompt-benchmark-v2.json"
    test_file_format: str | None = None
    output_dir: str = "results"
    output_file: str | None = None
    include_categories: list[str] = field(default_factory=list)
    exclude_categories: list[str] = field(default_factory=list)
    max_tests: int | None = None
    sleep_seconds: float = 0.0
    parallelism: int = 1
    requests_per_minute: float | None = None
    fail_threshold: float = 0.7
    stop_on_error: bool = False
    provider: ProviderConfig = field(default_factory=ProviderConfig)
    judge: JudgeConfig = field(default_factory=JudgeConfig)

    def resolved(self) -> "BenchmarkConfig":
        data = self.to_dict()
        data["prompt_file"] = _expand_path(self.prompt_file)
        data["test_file"] = _expand_path(self.test_file)
        data["output_dir"] = _expand_path(self.output_dir)
        if self.output_file:
            data["output_file"] = _expand_path(self.output_file)
        data["provider"]["api_key"] = (
            _expand_path(self.provider.api_key)
            if self.provider.api_key
            else self.provider.api_key
        )
        data["provider"]["base_url"] = (
            _expand_path(self.provider.base_url)
            if self.provider.base_url
            else self.provider.base_url
        )
        return benchmark_config_from_dict(data)

    def to_dict(self) -> dict[str, Any]:
        return asdict(self)


def benchmark_config_from_dict(data: dict[str, Any]) -> BenchmarkConfig:
    provider_data = data.get("provider", {})
    judge_data = data.get("judge", {})
    return BenchmarkConfig(
        prompt_file=data.get("prompt_file", BenchmarkConfig.prompt_file),
        test_file=data.get("test_file", BenchmarkConfig.test_file),
        test_file_format=data.get("test_file_format"),
        output_dir=data.get("output_dir", BenchmarkConfig.output_dir),
        output_file=data.get("output_file"),
        include_categories=list(data.get("include_categories", [])),
        exclude_categories=list(data.get("exclude_categories", [])),
        max_tests=data.get("max_tests"),
        sleep_seconds=float(data.get("sleep_seconds", 0.0)),
        parallelism=max(1, int(data.get("parallelism", 1))),
        requests_per_minute=float(data["requests_per_minute"])
        if data.get("requests_per_minute") is not None
        else None,
        fail_threshold=float(data.get("fail_threshold", 0.7)),
        stop_on_error=bool(data.get("stop_on_error", False)),
        provider=ProviderConfig(
            name=provider_data.get("name", ProviderConfig.name),
            model=provider_data.get("model"),
            embedding_model=provider_data.get("embedding_model"),
            rerank_model=provider_data.get("rerank_model"),
            api_key=provider_data.get("api_key"),
            api_key_env=provider_data.get("api_key_env"),
            base_url=provider_data.get("base_url"),
            api_version=provider_data.get("api_version"),
            aws_region=provider_data.get("aws_region"),
            project_id=provider_data.get("project_id"),
            location=provider_data.get("location"),
            headers=dict(provider_data.get("headers", {})),
            request_template=dict(provider_data.get("request_template", {})),
            response_text_path=provider_data.get("response_text_path"),
            response_tokens_path=provider_data.get("response_tokens_path"),
            timeout_seconds=float(provider_data.get("timeout_seconds", 60.0)),
            max_retries=int(provider_data.get("max_retries", 2)),
            retry_backoff_seconds=float(
                provider_data.get("retry_backoff_seconds", 1.5)
            ),
            temperature=float(provider_data.get("temperature", 0.2)),
            max_tokens=int(provider_data.get("max_tokens", 2048)),
        ),
        judge=JudgeConfig(
            strategy=judge_data.get("strategy", JudgeConfig.strategy),
            ollama_model=judge_data.get("ollama_model", JudgeConfig.ollama_model),
            timeout_seconds=float(judge_data.get("timeout_seconds", 30.0)),
            pass_threshold=float(judge_data.get("pass_threshold", 0.7)),
            review_threshold=float(judge_data.get("review_threshold", 0.4)),
            attacker_model=judge_data.get("attacker_model", "qwen3.5:9b"),
            attacker_base_url=judge_data.get(
                "attacker_base_url", "http://localhost:11434"
            ),
            attacker_api_key=judge_data.get("attacker_api_key"),
            attacker_api_key_env=judge_data.get("attacker_api_key_env"),
            attacker_temperature=float(judge_data.get("attacker_temperature", 0.3)),
            pattern_detectors_enabled=bool(
                judge_data.get("pattern_detectors_enabled", True)
            ),
            openai_moderation_enabled=bool(
                judge_data.get("openai_moderation_enabled", False)
            ),
            openai_moderation_model=judge_data.get(
                "openai_moderation_model", "omni-moderation-latest"
            ),
            openai_moderation_api_key=judge_data.get("openai_moderation_api_key"),
            openai_moderation_api_key_env=judge_data.get(
                "openai_moderation_api_key_env"
            ),
            openai_moderation_base_url=judge_data.get("openai_moderation_base_url"),
            perspective_enabled=bool(judge_data.get("perspective_enabled", False)),
            perspective_api_key=judge_data.get("perspective_api_key"),
            perspective_api_key_env=judge_data.get("perspective_api_key_env"),
            perspective_threshold=float(judge_data.get("perspective_threshold", 0.7)),
            harmjudge_enabled=bool(judge_data.get("harmjudge_enabled", False)),
            harmjudge_model=judge_data.get("harmjudge_model"),
            harmjudge_api_key=judge_data.get("harmjudge_api_key"),
            harmjudge_api_key_env=judge_data.get("harmjudge_api_key_env"),
            harmjudge_base_url=judge_data.get("harmjudge_base_url"),
            external_detector_url=judge_data.get("external_detector_url"),
            external_detector_api_key=judge_data.get("external_detector_api_key"),
            external_detector_api_key_env=judge_data.get(
                "external_detector_api_key_env"
            ),
            external_detector_headers=dict(
                judge_data.get("external_detector_headers", {})
            ),
            external_detector_timeout_seconds=float(
                judge_data.get("external_detector_timeout_seconds", 15.0)
            ),
            detector_weight=float(judge_data.get("detector_weight", 0.25)),
            detector_family_enabled={
                str(key): bool(value)
                for key, value in dict(
                    judge_data.get("detector_family_enabled", {})
                ).items()
            },
            detector_family_weights={
                str(key): float(value)
                for key, value in dict(
                    judge_data.get("detector_family_weights", {})
                ).items()
            },
        ),
    )


def load_benchmark_config(path: str) -> BenchmarkConfig:
    config_path = Path(_expand_path(path) or path)
    with config_path.open("r", encoding="utf-8") as handle:
        if config_path.suffix.lower() == ".json":
            raw = json.load(handle)
        else:
            if yaml is None:
                raise RuntimeError(
                    "PyYAML is required to load YAML config files. Install dependencies from requirements.txt."
                )
            raw = yaml.safe_load(handle) or {}
    return benchmark_config_from_dict(raw).resolved()


def save_benchmark_config(config: BenchmarkConfig, path: str) -> None:
    config_path = Path(path)
    config_path.parent.mkdir(parents=True, exist_ok=True)
    with config_path.open("w", encoding="utf-8") as handle:
        if config_path.suffix.lower() == ".json":
            json.dump(config.to_dict(), handle, indent=2, ensure_ascii=False)
        else:
            if yaml is None:
                raise RuntimeError(
                    "PyYAML is required to write YAML config files. Install dependencies from requirements.txt."
                )
            yaml.safe_dump(
                config.to_dict(), handle, sort_keys=False, allow_unicode=True
            )
