requirements.txt
openai>=1.54.0
anthropic>=0.18.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.71.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
httpx>=0.27.0
requests>=2.31.0
reportlab>=4.0.0
matplotlib>=3.7.0
kaleido>=0.2.1
PyYAML>=6.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
cohere>=5.13.0
boto3>=1.35.0
prometheus-client>=0.21.0
redis>=5.2.0
